feat: cross tab project sync
